+++ main-server/src/http_server/routes/v5/rp.js
@@ -82,6 +82,7 @@ router.post(
         },
         {
           synchronous: false,
+          validateIdentifier: false,
           apiVersion,
           ndidMemberAppType,
           ndidMemberAppVersion,